requirements.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
channels-redis
langchain
langgraph
langchain-google-genai